Convert a json file to parquet format

This function allows to convert a json or ndjson file to parquet format.

Two conversions possibilities are offered :

Convert to a single parquet file. Argument `path_to_parquet` must then be used;

Convert to a partitioned parquet file. Additionnal arguments `partition` and `partitioning` must then be used;

Usage

json_to_parquet(
  path_to_json,
  path_to_parquet,
  format = "json",
  partition = "no",
  ...
)

Arguments

path_to_json

string that indicates the path to the csv file

path_to_parquet

string that indicates the path to the directory where the parquet file will be stored

format

string that indicates if the format is "json" (by default) or "ndjson"

partition

string ("yes" or "no" - by default) that indicates whether you want to create a partitioned parquet file. If "yes", `"partitioning"` argument must be filled in. In this case, a folder will be created for each modality of the variable filled in `"partitioning"`.

...

additional format-specific arguments, see arrow::write_parquet() and arrow::write_dataset() for more informations.

Value

A parquet file, invisibly

Examples

# Conversion from a local json file to a single parquet file ::

json_to_parquet(
  path_to_json = system.file("extdata","iris.json",package = "parquetize"),
  path_to_parquet = tempdir()
)
#> Reading data...
#> Writing data...
#> ✔ The json file is available in parquet format under /tmp/RtmpME0hqS
#> Writing data...

# Conversion from a local ndjson file to a partitioned parquet file  ::

json_to_parquet(
  path_to_json = system.file("extdata","iris.ndjson",package = "parquetize"),
  path_to_parquet = tempdir(),
  format = "ndjson"
)
#> Reading data...
#> Writing data...
#> ✔ The ndjson file is available in parquet format under /tmp/RtmpME0hqS
#> Writing data...