Convert a sqlite file to parquet format

This function allows to convert a table from a sqlite file to parquet format.
The following extensions are supported : "db","sdb","sqlite","db3","s3db","sqlite3","sl3","db2","s2db","sqlite2","sl2".

Two conversions possibilities are offered :

Convert to a single parquet file. Argument `path_to_parquet` must then be used;

Convert to a partitioned parquet file. Additionnal arguments `partition` and `partitioning` must then be used;

Usage

sqlite_to_parquet(
  path_to_sqlite,
  table_in_sqlite,
  path_to_parquet,
  partition = "no",
  ...
)

Arguments

path_to_sqlite

string that indicates the path to the sqlite file

table_in_sqlite

string that indicates the name of the table to convert in the sqlite file

path_to_parquet

string that indicates the path to the directory where the parquet file will be stored

partition

string ("yes" or "no" - by default) that indicates whether you want to create a partitioned parquet file. If "yes", `"partitioning"` argument must be filled in. In this case, a folder will be created for each modality of the variable filled in `"partitioning"`.

...

additional format-specific arguments, see arrow::write_parquet() and arrow::write_dataset() for more informations.

Value

A parquet file, invisibly

Examples

# Conversion from a local sqlite file to a single parquet file :

sqlite_to_parquet(
  path_to_sqlite = system.file("extdata","iris.sqlite",package = "parquetize"),
  table_in_sqlite = "iris",
  path_to_parquet = tempdir()
)
#> Reading data...
#> Writing data...
#> ✔ The iris table from your sqlite file is available in parquet format under /tmp/RtmpC3WqGx
#> Writing data...

# Conversion from a local sqlite file to a partitioned parquet file  :

sqlite_to_parquet(
  path_to_sqlite = system.file("extdata","iris.sqlite",package = "parquetize"),
  table_in_sqlite = "iris",
  path_to_parquet = tempdir(),
  partition = "yes",
  partitioning =  c("Species")
)
#> Reading data...
#> Writing data...
#> ✔ The iris table from your sqlite file is available in parquet format under /tmp/RtmpC3WqGx
#> Writing data...