Convert a csv table to parquet format

This function allows to convert a csv table to parquet format.

csv_to_parquet(
  path_to_csv,
  url_to_csv,
  csv_as_a_zip = FALSE,
  filename_in_zip,
  path_to_parquet
)

Arguments

path_to_csv

string that indicates the path to the csv file

url_to_csv

string that indicates the URL of the csv file

csv_as_a_zip

boolean that indicates if the csv is stored in a zip

filename_in_zip

name of the csv file in the zip (useful if several csv are included in the zip). Required if `csv_as_a_zip` is TRUE.

path_to_parquet

string that indicates the path to the directory where the parquet file will be stored

Value

A parquet file

Details

Several conversion possibilities are offered :

From a locally stored file. The argument `path_to_csv` must then be used;

From a URL. The argument `url_to_csv` must then be used.

Examples

if (FALSE) {
# Conversion from a local csv file :

csv_to_parquet(
  path_to_csv = "Data/ac1.csv",
  path_to_parquet = "Data",
)

# Conversion from a URL and a csv file :

csv_to_parquet(
  url_to_csv = "https://stats.govt.nz/assets/Uploads/Research-and-development-survey/Research-and-development-survey-2021/Download-data/research-and-development-survey-2021-csv.csv",
  path_to_parquet = "Data",
)

# Conversion from a URL and a zipped file :

csv_to_parquet(
  url_to_csv = "https://www.insee.fr/fr/statistiques/fichier/3568617/equip-tour-transp-infra-2021.zip",
  csv_as_a_zip = TRUE,
  filename_in_zip = "equip-tour-transp-infra-2021.csv",
  path_to_parquet = "Data",
)
}