Convert a csv file to parquet format

This function allows to convert a csv file to parquet format.

Usage

csv_to_parquet(
  path_to_csv,
  url_to_csv,
  csv_as_a_zip = FALSE,
  filename_in_zip,
  path_to_parquet,
  compression = "snappy",
  compression_level = NULL,
  partition = "no",
  encoding = "UTF-8",
  ...
)

Arguments

path_to_csv

string that indicates the path to the csv file

url_to_csv

string that indicates the URL of the csv file

csv_as_a_zip

boolean that indicates if the csv is stored in a zip

filename_in_zip

name of the csv file in the zip (useful if several csv are included in the zip). Required if `csv_as_a_zip` is TRUE.

path_to_parquet

string that indicates the path to the directory where the parquet file will be stored

compression

compression algorithm. Default "snappy".

compression_level

compression level. Meaning depends on compression algorithm.

partition

string ("yes" or "no" - by default) that indicates whether you want to create a partitioned parquet file. If "yes", `"partitioning"` argument must be filled in. In this case, a folder will be created for each modality of the variable filled in `"partitioning"`.

encoding

string that indicates the character encoding for the input file.

...

additional format-specific arguments, see arrow::write_parquet() and arrow::write_dataset() for more informations.

Value

A parquet file, invisibly

Details

Several conversion possibilities are offered :

From a locally stored file. Argument `path_to_csv` must then be used;

From a URL. Argument `url_to_csv` must then be used.

Two conversions possibilities are offered :

Convert to a single parquet file. Argument `path_to_parquet` must then be used;

Convert to a partitioned parquet file. Additionnal arguments `partition` and `partitioning` must then be used;

Note

Be careful, if the zip size exceeds 4 GB, the function may truncate the data (because unzip() won't work reliably in this case - see here). In this case, it's advised to unzip your csv file by hand (for example with 7-Zip) then use the function with the argument `path_to_csv`.

Examples

# Conversion from a local csv file to a single parquet file :

csv_to_parquet(
  path_to_csv = parquetize_example("region_2022.csv"),
  path_to_parquet = tempdir()
)
#> Reading data...
#> Writing data...
#> ✔ The csv file is available in parquet format under /tmp/RtmpNKHnY7
#> Writing data...

# Conversion from a local csv file  to a partitioned parquet file  :

csv_to_parquet(
  path_to_csv = parquetize_example("region_2022.csv"),
  path_to_parquet = tempdir(),
  partition = "yes",
  partitioning =  c("REG")
)
#> Reading data...
#> Writing data...
#> ✔ The csv file is available in parquet format under /tmp/RtmpNKHnY7
#> Writing data...

# Conversion from a URL and a csv file with "gzip" compression :

csv_to_parquet(
  url_to_csv = "https://stats.govt.nz/assets/Uploads/Research-and-development-survey/Research-and-development-survey-2021/Download-data/research-and-development-survey-2021-csv.csv",
  path_to_parquet = tempdir(),
  compression = "gzip",
  compression_level = 5
)
#> Reading data...
#> ✔ The csv file is available in parquet format under /tmp/RtmpNKHnY7
#> Reading data...

# Conversion from a URL and a zipped file :

csv_to_parquet(
  url_to_csv = "https://www.insee.fr/fr/statistiques/fichier/5057840/cog_ensemble_2021_csv.zip",
  csv_as_a_zip = TRUE,
  filename_in_zip = "commune2021.csv",
  path_to_parquet = tempdir()
)
#> Reading data...
#> Error in curl_download(url_to_csv, tempfile()): Timeout was reached: [] Failed to connect to www.insee.fr port 443 after 7568 ms: Connection timed out